Things reasonably fixed up.
Cut the number of lookaheads to 3, which seems to produce reasonable
puzzles.  Hard to really generate a specific puzzle on demand.

I don't know what changed, but for some reason coverage=25 never seems
to produce a solvable puzzle.

Updated the fullpuzzle to include a week of puzzles from the Merc.
Dumping more statistics